A major revision, use EMLassembly line
The previous commits of this repository used the R functions coded similar to the Postgres-to-EML workflow. This version, I worked with EDI to recode all the functions and this is a bridge code to connect EMLassemblyline. We hope that EDI will continue to maintain the EMLassembly line and reassume any major change done on the EML package will not break this bridge code.